fix(ui): validate PR chip identity and coverage boundary
